make createGlobalStyle strict-mode compliant
