+++ packages/styled-components/src/constructors/createGlobalStyle.js
@@ -1,16 +1,14 @@
 // @flow
-import React, { useContext, useEffect, useRef } from 'react';
+import React, { useContext, useLayoutEffect, useRef } from 'react';
 import { STATIC_EXECUTION_CONTEXT } from '../constants';
 import GlobalStyle from '../models/GlobalStyle';
 import { useStyleSheet, useStylis } from '../models/StyleSheetManager';
+import { ThemeContext } from '../models/ThemeProvider';
+import type { Interpolation } from '../types';
 import { checkDynamicCreation } from '../utils/checkDynamicCreation';
 import determineTheme from '../utils/determineTheme';
-import { ThemeContext } from '../models/ThemeProvider';
-import { EMPTY_ARRAY } from '../utils/empties';
 import generateComponentId from '../utils/generateComponentId';
 import css from './css';
-
-import type { Interpolation } from '../types';
 
 type GlobalStyleComponentPropsType = Object;
 
@@ -51,6 +49,22 @@ export default function createGlobalStyle(
       );
     }
 
+    if (typeof window === 'undefined') {
+      renderStyles(instance, props, styleSheet, theme, stylis);
+    } else {
+      // this conditional is fine because it is compiled away for the relevant builds during minification,
+      // resulting in a single unguarded hook call
+      // eslint-disable-next-line react-hooks/rules-of-hooks
+      useLayoutEffect(() => {
+        renderStyles(instance, props, styleSheet, theme, stylis);
+        return () => globalStyle.removeStyles(instance, styleSheet);
+      }, [instance, props, styleSheet, theme, stylis]);
+    }
+
+    return null;
+  }
+
+  function renderStyles(instance, props, styleSheet, theme, stylis) {
     if (globalStyle.isStatic) {
       globalStyle.renderStyles(instance, STATIC_EXECUTION_CONTEXT, styleSheet, stylis);
     } else {
@@ -61,10 +75,6 @@ export default function createGlobalStyle(
 
       globalStyle.renderStyles(instance, context, styleSheet, stylis);
     }
-
-    useEffect(() => () => globalStyle.removeStyles(instance, styleSheet), EMPTY_ARRAY);
-
-    return null;
   }
 
   // $FlowFixMe
